pyproject.toml
[tool.poetry]
name = "slide-captcha"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.0.0"
httpx = "^0.25.0"
loguru = "^0.7.2"
ultralytics = "^8.0.178"
onnx = "^1.14.1"
onnxruntime = "^1.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"